Implement proper HDR image capture, and bump to iOS 14.1
I've implemented this by extracting the gain map from the raw image data through CGImageSourceCopyAuxiliaryDataInfoAtIndex(), with the key kCGImageAuxiliaryDataTypeHDRGainMap. The benefits of this approach is that I can capture proper HDR images with gain maps, and the proper colorspace (Display P3). The downside is that the gain map extraction key only came with iOS 14.1. I chose to bump the minimum, as most users are likely going to be running a later version anyway.