Finally fixed that focus slider crash!
The issue is that I was incorrectly assuming that Apple wouldn't have shipped a fixed-position ultra wide camera in 2019 to 2023. However, they did. Thanks to the same friend for loaning me an iPhone 11 to use and finally fix this one bug that has existed for a while.
Also fixed some dialogs.